+++ packages/core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @xnetjs/core
 
+## 0.0.3
+
 ## 0.0.2
 
 ### Patch Changes