+++ packages/crypto/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @xnetjs/crypto
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xnetjs/core@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes