+++ packages/history/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @xnetjs/history
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @xnetjs/data@0.0.3
+  - @xnetjs/sync@0.0.3
+  - @xnetjs/core@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes