+++ packages/react/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @xnetjs/react
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies [[`6183829`](https://github.com/crs48/xNet/commit/618382920002a39f00e4f5f4a2ae604c2aef4fa6)]:
+  - @xnetjs/billing@0.0.2
+  - @xnetjs/plugins@0.0.3
+  - @xnetjs/runtime@0.0.2
+  - @xnetjs/history@0.0.3
+  - @xnetjs/data-bridge@0.0.3
+  - @xnetjs/data@0.0.3
+  - @xnetjs/sync@0.0.3
+  - @xnetjs/identity@0.0.3
+  - @xnetjs/crypto@0.0.3
+  - @xnetjs/core@0.0.3
+
 ## 0.0.2
 
 ### Patch Changes