+++ packages/sqlite/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @xnetjs/sqlite
 
+## 0.0.3
+
 ## 0.0.2
 
 ### Patch Changes